backend/requirements.txt
Django
djangorestframework
django-smtp-ssl
django-cors-headers
python-decouple
mssql-django
pyjwt
psycopg2


# Django==4.0.8
# djangorestframework==3.14.0
# django-smtp-ssl==1.0
# django-cors-headers==3.13.0
# python-decouple==3.6
# mssql-django==1.1.3
# pyjwt==2.6.0
# psycopg2==2.9.5